make time interval its own component
